+++ fsw/src/bpnode_adu_in.h
@@ -129,7 +129,7 @@ void BPNode_AduIn_AppMain(void);
  *
  *  \param[in] ChanId Channel ID for this task
  */
-void BPNode_AduIn_TaskExit(uint8 ChanId);
+void BPNode_AduIn_TaskExit(uint32 ChanId);
 
 
 #endif /* BPNODE_ADU_IN_H */
